move openRuleManagementPageViaBreadcrumbs to rule management file
