+++ x-pack/test/security_solution_cypress/cypress/e2e/detection_response/rule_creation/custom_query_rule_data_view.cy.ts
@@ -63,7 +63,8 @@ import {
 } from '../../../tasks/create_new_rule';
 
 import { login } from '../../../tasks/login';
-import { openRuleManagementPageViaBreadcrumbs, visit } from '../../../tasks/navigation';
+import { visit } from '../../../tasks/navigation';
+import { openRuleManagementPageViaBreadcrumbs } from '../../../tasks/rules_management';
 import { getDetails, waitForTheRuleToBeExecuted } from '../../../tasks/rule_details';
 
 import { CREATE_RULE_URL } from '../../../urls/navigation';
