+++ x-pack/test/security_solution_cypress/cypress/e2e/detection_response/rule_creation/machine_learning_rule.cy.ts
@@ -50,7 +50,8 @@ import {
   selectMachineLearningRuleType,
 } from '../../../tasks/create_new_rule';
 import { login } from '../../../tasks/login';
-import { openRuleManagementPageViaBreadcrumbs, visit } from '../../../tasks/navigation';
+import { visit } from '../../../tasks/navigation';
+import { openRuleManagementPageViaBreadcrumbs } from '../../../tasks/rules_management';
 import { CREATE_RULE_URL } from '../../../urls/navigation';
 
 // TODO: https://github.com/elastic/kibana/issues/161539
