+++ x-pack/test/security_solution_cypress/cypress/e2e/detection_response/rule_creation/override.cy.ts
@@ -60,7 +60,7 @@ import { login } from '../../../tasks/login';
 import { visit } from '../../../tasks/navigation';
 import { getDetails, waitForTheRuleToBeExecuted } from '../../../tasks/rule_details';
 import { CREATE_RULE_URL } from '../../../urls/navigation';
-import { openRuleManagementPageViaBreadcrumbs } from '../../../tasks/navigation';
+import { openRuleManagementPageViaBreadcrumbs } from '../../../tasks/rules_management';
 
 // TODO: https://github.com/elastic/kibana/issues/161539
 describe('Rules override', { tags: ['@ess', '@serverless', '@brokenInServerless'] }, () => {
