+++ x-pack/test/security_solution_cypress/cypress/tasks/navigation.ts
@@ -10,7 +10,6 @@ import { encode } from '@kbn/rison';
 import type { ROLES } from '@kbn/security-solution-plugin/common/test';
 import { NEW_FEATURES_TOUR_STORAGE_KEYS } from '@kbn/security-solution-plugin/common/constants';
 import { hostDetailsUrl, userDetailsUrl } from '../urls/navigation';
-import { LAST_BREADCRUMB, RULE_MANAGEMENT_PAGE_BREADCRUMB } from '../screens/breadcrumbs';
 import { constructUrlWithUser, getUrlWithRoute, User } from './login';
 
 export const visit = (
@@ -91,12 +90,6 @@ export const visitHostDetailsPage = (hostName = 'suricata-iowa') => {
 export const visitUserDetailsPage = (userName = 'test') => {
   visitWithTimeRange(userDetailsUrl(userName));
 };
-
-export function openRuleManagementPageViaBreadcrumbs(): void {
-  cy.log('Navigate back to rules table via breadcrumbs');
-  cy.get(`${RULE_MANAGEMENT_PAGE_BREADCRUMB}:not(${LAST_BREADCRUMB})`).click();
-  cy.get(`${RULE_MANAGEMENT_PAGE_BREADCRUMB}${LAST_BREADCRUMB}`).should('exist');
-}
 
 /**
  * For all the new features tours we show in the app, this method disables them
