+++ x-pack/test/security_solution_cypress/cypress/tasks/rules_management.ts
@@ -6,6 +6,7 @@
  */
 
 import type { ROLES } from '@kbn/security-solution-plugin/common/test';
+import { LAST_BREADCRUMB, RULE_MANAGEMENT_PAGE_BREADCRUMB } from '../screens/breadcrumbs';
 import { RULES_MANAGEMENT_URL } from '../urls/rules_management';
 import { resetRulesTableState } from './common';
 import { visit } from './navigation';
@@ -13,4 +14,10 @@ import { visit } from './navigation';
 export function visitRulesManagementTable(role?: ROLES): void {
   resetRulesTableState(); // Clear persistent rules filter data before page loading
   visit(RULES_MANAGEMENT_URL, { role });
+}
+
+export function openRuleManagementPageViaBreadcrumbs(): void {
+  cy.log('Navigate back to rules table via breadcrumbs');
+  cy.get(`${RULE_MANAGEMENT_PAGE_BREADCRUMB}:not(${LAST_BREADCRUMB})`).click();
+  cy.get(`${RULE_MANAGEMENT_PAGE_BREADCRUMB}${LAST_BREADCRUMB}`).should('exist');
 }
